% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{AddCsvColToMeta}
\alias{AddCsvColToMeta}
\title{Updates meta csv column}
\usage{
AddCsvColToMeta()
}
\description{
Updates meta csv column
}
\details{
In previous versions of datatrack, whether to save as csv was specified in the config file
There was no 'csv' column in the metadata. Although this saved a column, it required more configuration
This function adds the 'csv' column and populates the rows according to the json config
}

