% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{D3Inspector}
\alias{D3Inspector}
\title{Main function to generate the viewer}
\usage{
D3Inspector(group = FALSE)
}
\arguments{
\item{group}{string; name to filter nodes by}
}
\description{
Main function to generate the viewer
}
\details{
Gets the metadata in the appropriate json format then passes it to the dataGraph function of the dviewer package
}

