% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{Example1}
\alias{Example1}
\title{Creates a few example data objects with some dependencies between them}
\usage{
Example1()
}
\description{
Creates a few example data objects with some dependencies between them
}

