% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{Example2}
\alias{Example2}
\title{Example showing attempting some incompatible dependencies}
\usage{
Example2()
}
\description{
Example showing attempting some incompatible dependencies
}
\details{
This should be done after setting up the initial data objects with Example1
Here, we try to save a data object that has indirect dependencies on different versions of the same
name. This should not be allowed, but currently it is.
}

