% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{Example5}
\alias{Example5}
\title{Example showing reading data specifying 2 names}
\usage{
Example5()
}
\description{
Example showing reading data specifying 2 names
}
\details{
When 2 or more names are specified, it will return the last accessed version of any of those names
}

