% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{Example7}
\alias{Example7}
\title{Example showing saving with annotations}
\usage{
Example7()
}
\description{
Example showing saving with annotations
}
\details{
Annotations can either be a list or just a string. This example saves one of each
}

