% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{ExampleA1}
\alias{ExampleA1}
\title{Creates a few example data objects with some dependencies between them}
\usage{
ExampleA1()
}
\description{
Creates a few example data objects with some dependencies between them
}

