% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{FixMeta}
\alias{FixMeta}
\title{Fix the structure of the metadata dataframe}
\usage{
FixMeta(meta)
}
\description{
Restores the structure of the metadata dataframe, which might be
necessary if, during a change to datatrack e.g. a new version of the package
extra columns are added, removed or reordered
}
\details{
An empty correct meta dataframe is created, then each of the existing
columns are attempted to be matched to a correct column. If no match for an existing
column is found, it is discarded with a warning. If a correct column has not had a match found
for it, it is left blank.

In specific cases of upgrades, transformations might need to be made to the existing columns. If this
happens this function should be added to appropriately. For example, if csv column is removed an a 'class' column
is addeed instead, then specific tranformation from existing csv = 1 rows to class = 'csv' needs to be coded.
}

