% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{GetMeta}
\alias{GetMeta}
\title{Returns the metadata row for the given name/version pair}
\usage{
GetMeta(name, version)
}
\arguments{
\item{name}{string}

\item{version}{int}
}
\value{
list
}
\description{
Returns the metadata row for the given name/version pair
}
\details{
Reads the metadata, filters to the correct row, converts to list,
then converts json encoded values to lists
}
