% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{Inspector}
\alias{Inspector}
\title{Display the visual datatrack inspector}
\usage{
Inspector(group = FALSE)
}
\arguments{
\item{group}{string; name of the datatrack data object to filter nodes by.
Will only show nodes that are dependent on, or a dependency of an version of this name}
}
\description{
Generate the visualisation of saved data objects, their parameters and dependencies
}
\details{
Gets the metadata in the appropriate json format then passes it to the dataGraph function of the dviewer package
}

