% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publicreadwrite.R
\name{ReadDataobject}
\alias{ReadDataobject}
\title{Read a Dataobject of a Given Name}
\usage{
ReadDataobject(name = NULL, purpose = NA, include.meta = TRUE,
  params = NULL, dependencies = NULL, false.if.missing = FALSE,
  optional = FALSE, use.last.accessed = TRUE, version = NULL)
}
\arguments{
\item{name}{String; Optional. The dataobject to read, eg "clusters", "features" etc.  If ommited, will prompt the user with a choice of available names}

\item{purpose}{string; Displayed to the user when prompting}

\item{include.meta}{bool; if true, will wrap the data to return in a list that also contains the metadata}

\item{params}{list; Optional. If supplied, will only consider returning a dataobject with the matching params}

\item{dependencies}{list; optional. If supplied, will only consider returning a dataobject with the matching}

\item{false.if.missing}{bool; if true, will return false if the file is missing}

\item{optional}{boolean; if true, an option is added to select none and return false}

\item{use.last.accessed}{boolean; if true (default) will look for the version that was last written or read in this session}

\item{version}{int; If the required version is known, then it can be supplied. This version will be used if it exists}
}
\value{
if include.meta, will return a list that has a 'data' key, that contains the data read in if include.meta is false, will return the data read in, eg data frame if it's a csv
}
\description{
Given the name and optionally the parameters and dependencies, will read the dataobject previously saved in a datatrack project.
}
\details{
1) if name is not supplied, will ask user,
   2) then if version is supplied, find the that version of the dataobject name. If it doesn't exist, will return false or stop (depending on false.if.missing)
   3) looks for the last accessed version if it matches the version OR  params, dependencies
   4) if no matching last accessed verion is found, it will ask the user to select a version
      This means that for a particular set of params and dependencies, only one version of a particular dataobject name can be used within the same run
   5) if the file is found, then it will set the 'last accessed' flag
      on the chosen dataobject and it dependencies
}

