% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{ReadMeta}
\alias{ReadMeta}
\title{Reads the csv of metadata from disk}
\usage{
ReadMeta()
}
\description{
Reads the csv of metadata from disk
}
\details{
If the file doesn't exist, creates an empty one with the correct columns
}

