% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{ReadOutput}
\alias{ReadOutput}
\title{Legacy Wrapper for ReadDataobject}
\usage{
ReadOutput(name = NULL, purpose = NA, include.meta = TRUE,
  params = NULL, dependencies = NULL, false.if.missing = FALSE,
  optional = FALSE, use.last.accessed = TRUE, version = NULL)
}
\description{
Legacy Wrapper for ReadDataobject
}

