% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{RunAllExamples}
\alias{RunAllExamples}
\title{Runs all examples
This should produce consistent set of dataobjects and metadata}
\usage{
RunAllExamples()
}
\description{
Runs all examples
This should produce consistent set of dataobjects and metadata
}

