% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{SetConfig}
\alias{SetConfig}
\title{sets the config values and saves the config file}
\usage{
SetConfig(...)
}
\arguments{
\item{...}{name value pairs as arguments or a list of name value pairs}
}
\description{
sets the config values and saves the config file
}

