% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{SetConfigFile}
\alias{SetConfigFile}
\title{Sets the path to the config json file}
\usage{
SetConfigFile(path)
}
\arguments{
\item{path}{string}
}
\description{
Sets the path to the config json file
}
\details{
By Default, datatrack will look for a file called 'datatrack.config.json' in the working directory
This can be overridden by this function. This is not recommended during normal use of datatrack,
since it needs to be called after datatrack is loaded but before it is used, leaving it prone to error.
It is designed for use for unit testing so that datatrack projects can be created and destroyed
}

