% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{WriteOutput}
\alias{WriteOutput}
\title{Legacy Wrapper for WriteDataobject}
\usage{
WriteOutput(x, name, params = list(), dependencies = list(),
  check.before.overwrite = TRUE)
}
\description{
Legacy Wrapper for WriteDataobject
}

