% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{WriteStructuredOutput}
\alias{WriteStructuredOutput}
\title{Legacy Wrapper for WriteStructuredDataobject}
\usage{
WriteStructuredOutput(x, check.before.overwrite = TRUE)
}
\description{
Legacy Wrapper for WriteStructuredDataobject
}

