% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accesslog.R
\name{ClearAccessLog}
\alias{ClearAccessLog}
\title{Clear the datatrack access log}
\usage{
ClearAccessLog()
}
\description{
Clear the access log used to save the most recently accessed version of each name
}

