% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{GetChecksum}
\alias{GetChecksum}
\title{Compute a Checksum of the Metadata}
\usage{
GetChecksum(ignore.cols = c("date", "callstack", "system"), algo = "sha1")
}
\description{
To quickly compare if the metadata of 2 datatrack projects are the same
We can compute a checksum which ignores the date column.
}
\details{
After removing the columns that don't matter for meta comparisons, 
e.g date, system, calstack column of the metadata, performs a hash of the dataframe and returns is
}

