% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{ReadOutput}
\alias{ReadOutput}
\title{Legacy Wrapper for ReadDataobject}
\usage{
ReadOutput(name = NULL, purpose = NA, include.meta = TRUE,
  params = NULL, dependencies = NULL, false.if.missing = FALSE,
  optional = FALSE, use.last.accessed = TRUE, version = NULL)
}
\arguments{
\item{name}{character}

\item{purpose}{character}

\item{include.meta}{boolean}

\item{params}{list}

\item{dependencies}{list}

\item{false.if.missing}{boolean}

\item{optional}{boolean}

\item{use.last.accessed}{boolean}

\item{version}{int}
}
\value{
list
}
\description{
Legacy Wrapper for ReadDataobject
}

