% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publicreadwrite.R
\name{WriteDataobject}
\alias{WriteDataobject}
\title{Write a Dataobject to the Datatrack Project}
\usage{
WriteDataobject(x, name, params = list(), dependencies = list(),
  check.before.overwrite = TRUE, annotations = list())
}
\arguments{
\item{x}{data.frame or Object}

\item{name}{string what to call the dataobject. This is used in: the filename and the meta file to keep track of versions and dependencies}

\item{params}{list The parameters used when generating this dataobject. Saved in the meta file.}

\item{dependencies}{list The dataobjects created from previous steps that were used as input data for the process that created this dataobject List of name/version pairs}

\item{check.before.overwrite}{boolean if false, overwrite an existing dataobject if the given parameters and dependencies are the same as the existing one
If the dependency version is given as 0, then datatrack will attempt to look up the last accessed version of that name to use. In other words, use 0 to
specify that datatrack should use the last accessed version as the dependency}
}
\value{
the version that it gets saved as
}
\description{
Given any R object, along with parameters and dependencies used when creating the value, will
save the object to disk and record the metadata in the datatrack project
}
\details{
dataobject will be saved with the filename like: name.version.csv  The version is detected automatically.
If dataobject for this name with the same parameters and dependencies is already saved, then it will be overwritten (after user confirmation).
If params or dependencies are different, then a new version is saved. The version number is created automatically.

The version for the dependency can be ommited (by putting a value < 1). If this happens, then it will be assumed to be the last accessed version
of the dependency dataobject. So, the process which generated the dataobject will have accesed some other dataobject it depends on. Then, when it saves its dataobject
it need only pass the name of the input dataobject without the version, and the function will know which version it was.
}

