% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{WriteOutput}
\alias{WriteOutput}
\title{Legacy Wrapper for WriteDataobject}
\usage{
WriteOutput(x, name, params = list(), dependencies = list(),
  check.before.overwrite = TRUE)
}
\arguments{
\item{x}{mixed}

\item{name}{character}

\item{params}{list}

\item{dependencies}{list}

\item{check.before.overwrite}{boolean}
}
\value{
int the version of the name that was written
}
\description{
Legacy Wrapper for WriteDataobject
}

