% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{WriteStructuredDataobject}
\alias{WriteStructuredDataobject}
\title{writes a dataobject where the meta values for the dataobject are included in the list
rather than passed as separate parameters}
\usage{
WriteStructuredDataobject(x, check.before.overwrite = TRUE)
}
\arguments{
\item{x}{mixed the data to write}

\item{check.before.overwrite}{boolean whether to prompt for user confirmation before overwriting an existing version}
}
\description{
writes a dataobject where the meta values for the dataobject are included in the list
rather than passed as separate parameters
}
\details{
Simply takes the parameters from the list and uses them for WriteDataobject
}

